Gave up, started Map
I've accepted my losses with the sets, so I have moved on to start implementing the maps. No matter what I do, I cannot seem to get past the segfaults in the sets. I have probably spent 5-6 hours just debugging setfaults and it this point it is a lost cause. Starting from scratch for maps to see if maybe I can make it potentially work half as expected.